backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
Pillow
transformers
torch
